Move SelectPreviewDevice to admin/components
