+++ front/app/components/admin/SelectPreviewDevice/index.tsx
